Исправления в платформе Кибермишки (#131)
* fix matrixPicture

* fix type cast

* fix matrixAnimation

* sync platform with client

* add timer for click method

* sync platform with client

* fix type for Sound and Note components

* fix MatrixAnimation

* fix typo

* fix matrixAnimation and btn isClicked

* fix pwm speed (via matrix frequency)

* note in progress

* update component speakerNote

* add note

* add sounds. Sound note component ready

+++ compiler/library/blg-mb-1-a7/1.0/source/SpeakerSound.hpp
@@ -27,9 +27,11 @@ public:
     }
 
     // duration in ms
-    void play(Sound *sound, const uint16_t duration) {
+    void play(Sound *sound, const uint32_t duration) {
 
         mrx::hal::speaker::startSound(sound, duration);
+
+        // mrx::hal::speaker::startSound(&LaughterSound, duration);
     }
 
     void stop() {
